Refactor for integration with linsights
Word, Words, Sentence and Sentences classes have been refactored to add
factory methods that instantiate objects of the respective class based
on the context. For instance, a Word object may be created from a data
file containing pre-generated features using the from_line factory
method and a Word object may be created from (word, root, pos, chunk)
object using the from_group factory method.

The difference in approach to create objects is needed because the
meaning of a Word is different when instantiated to represent a word
from pre-built training data file from that instantiated to represent a
word that should be classified into an appropriate uncertainty class.