Modify classify() to work with linsights
The classify() function now takes two types of data to classify: a
sentence in plain text and a group of (word, root, pos, chunk). As of
this change, the feature to classify sentence is not implemented as it
requires the code to convert the plain text sentence into group of
(word, root, pos, chunk). The return type of the response from
classify() can either be a list (in case of cue-level classification) or
a scalar (in case of sentence-level classification).