Add wrapper class for the uncertainty classifier
The wrapper class simplifies the implementation of the uncertainty
classifier and improves performance. The earlier, function-based,
approach to prediction using the classifier required the classifier to
be loaded from disk for each call to classify. However, with the
class-based implementation, the classifier can be loaded once and used
multiple times to predict.