Require subcommand to be specified when run
Invocation of main.py file is meaningless unless an appopriate command
is specified. The change makes the command a required attribute.
Additionally, the import of the classifier is now moved to after the
command line arguments are parsed to avoid incuring the overhead of
loading modules in case the command line arguments are not specified
appropriately.