Refresh vectorizer and models with updated data
The data used to train the vectorizer and the models used the stem, part
of speech tag and chunk tag as-is from the data files provided by
Vincze. To achieve consistency, the training data files have been
updated by using the stemmer, part of speech tagger and chunk tagger
from the linsights project (andymeneely/sira-nlp).