Train classifiers using the original training data
The tokens in the original training data files were associated with one
of the following five uncertainty classes:

 * O for Other or Certain.
 * (B-/I-)speculation_modal_probable_ for Epistemic.
 * (B-/I-)speculation_hypo_doxastic _ for Doxastic.
 * (B-/I-)speculation_hypo_condition _ for Condition.
 * (B-/I-)speculation_hypo_investigation _ for Investigation.

Prior to this change, the training data files were modified to map the
original uncertainty class to a single character for simplicity. With
this change, however, the source code that trains the classifier no
longer expects the training data files to be modified and handles the
mapping of uncertainty class to corresponding single character
transparently.

Additionally, the use of csv module to read TSV training data files was
causing problems as the fields in the original training data files were
not appropriately quoted. As a result, custom functions---read_tsv and
write_tsv---from uncertainty.helpers is being used to read and write TSV
files.